# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BeautyBook API
  version: 1.0.0
  description: |
    Especificacao OpenAPI da API REST do BeautyBook, baseada nos epicos, user stories
    e regras de negocio do projeto.

    Principais regras refletidas nesta API:
    - Autenticacao obrigatoria para acesso aos recursos protegidos
    - Controle de acesso por perfil (`ADMIN` e `CLIENT`)
    - Gestao de servicos com regras condicionais para `necessitaAvaliacao`
    - Configuracao de expediente, dias de funcionamento e intervalos
    - Consulta de disponibilidade considerando agenda, sobreposicao e duracao do servico
    - Criacao e cancelamento de agendamentos com regras operacionais do salao
servers:
  - url: http://localhost:3000/api
    description: Ambiente local
...
security:
  - bearerAuth: []
paths:
...
  /services:
    get:
      tags: [Servicos]
      summary: Lista servicos
      description: Lista o catalogo de servicos disponiveis para usuarios autenticados.
      responses:
        '200':
          description: Lista de servicos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Servicos]
      summary: Cadastra um servico
      description: Recurso restrito ao perfil `ADMIN`.
      x-roles:
        - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreateRequest'
      responses:
        '201':
          description: Servico criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          description: Dados invalidos para cadastro do servico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                invalidTimeFormat:
                  summary: Hora fora do formato esperado
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: tempoServico
                        message: Formato de hora invalido. Use HH:MM
                invalidTimeRange:
                  summary: tempoServico fora da faixa permitida
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: tempoServico
                        message: tempoServico deve estar entre 00:15 e 08:00
                invalidValue:
                  summary: Valor menor ou igual a zero
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: valor
                        message: valor deve ser maior que zero
                missingRequiredFields:
                  summary: Campos obrigatorios ausentes quando nao exige avaliacao
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: tempoServico
                        message: tempoServico e obrigatorio quando necessitaAvaliacao for false
                      - field: valor
                        message: valor e obrigatorio quando necessitaAvaliacao for false
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Nome de servico duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: SERVICE_NAME_ALREADY_EXISTS
                message: Ja existe um servico com este nome
  /services/{nomeServico}:
    get:
      tags: [Servicos]
      summary: Busca um servico por nome
      parameters:
        - $ref: '#/components/parameters/NomeServico'
      responses:
        '200':
          description: Servico encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Servicos]
      summary: Atualiza um servico
      description: Recurso restrito ao perfil `ADMIN`.
      x-roles:
        - ADMIN
      parameters:
        - $ref: '#/components/parameters/NomeServico'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '200':
          description: Servico atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          description: Dados invalidos para atualizacao do servico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                invalidTimeFormat:
                  summary: Hora fora do formato esperado
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: tempoServico
                        message: Formato de hora invalido. Use HH:MM
                invalidTimeRange:
                  summary: tempoServico fora da faixa permitida
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: tempoServico
                        message: tempoServico deve estar entre 00:15 e 08:00
                invalidValue:
                  summary: Valor menor ou igual a zero
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: valor
                        message: valor deve ser maior que zero
                missingRequiredFields:
                  summary: Campos obrigatorios ausentes quando nao exige avaliacao
                  value:
                    code: VALIDATION_ERROR
                    message: Dados do servico invalidos
                    details:
                      - field: tempoServico
                        message: tempoServico e obrigatorio quando necessitaAvaliacao for false
                      - field: valor
                        message: valor e obrigatorio quando necessitaAvaliacao for false
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Nome de servico duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Servicos]
      summary: Exclui um servico
      description: Recurso restrito ao perfil `ADMIN`.
      x-roles:
        - ADMIN
      parameters:
        - $ref: '#/components/parameters/NomeServico'
      responses:
        '204':
          description: Servico excluido com sucesso
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    NomeServico:
      name: nomeServico
      in: path
      required: true
      schema:
        type: string
      description: Nome do servico na URL. Quando houver espacos, deve ser enviado com URL encoding.
...
  responses:
    UnauthorizedError:
      description: Usuario nao autenticado ou token invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Autenticacao obrigatoria
    ForbiddenError:
      description: Usuario autenticado sem permissao para o recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Voce nao possui permissao para acessar este recurso
    NotFoundError:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_NOT_FOUND
            message: Recurso nao encontrado
...
  schemas:
...
    Service:
      allOf:
        - $ref: '#/components/schemas/ServiceBase'
        - type: object
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    ServiceBase:
      type: object
      required:
        - name
        - necessitaAvaliacao
      properties:
        name:
          type: string
          maxLength: 44
          example: Corte feminino
        description:
          type: string
          nullable: true
          example: Corte com finalizacao simples
        necessitaAvaliacao:
          type: boolean
          example: false
        tempoServico:
          allOf:
            - $ref: '#/components/schemas/LocalTime'
          nullable: true
          description: Obrigatorio quando `necessitaAvaliacao = false`. Faixa permitida de `00:15` a `08:00`.
          example: '01:30'
        valor:
          type: number
          format: double
          nullable: true
          minimum: 0.01
          description: Obrigatorio quando `necessitaAvaliacao = false`.
          example: 120.0
    ServiceCreateRequest:
      allOf:
        - $ref: '#/components/schemas/ServiceBase'
    ServiceUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/ServiceBase'
...
    LocalTime:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: '13:30'
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BUSINESS_RULE_VIOLATION
        message:
          type: string
          example: Nao foi possivel concluir a operacao
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: '#/components/schemas/FieldValidationError'
    FieldValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: tempoServico
        message:
          type: string
          example: Formato de hora invalido. Use HH:MM